+++ resolve/vector/VectorHandlerCuda.cpp
@@ -318,7 +318,7 @@ namespace ReSolve
   }
 
   /**
-   * @brief Scale a vector by a diagonal matrix in HIP
+   * @brief Scale a vector by a diagonal matrix in CUDA
    *
    * @param[in]  diag - vector representing the diagonal matrix
    * @param[in, out]  vec - vector to be scaled
@@ -333,6 +333,29 @@ namespace ReSolve
   void VectorHandlerCuda::scal(vector::Vector* diag, vector::Vector* vec)
   {
     real_type* diag_data = diag->getData(memory::DEVICE);
+    real_type* vec_data  = vec->getData(memory::DEVICE);
+    index_type n         = vec->getSize();
+    cuda::scale(n, diag_data, vec_data);
+    vec->setDataUpdated(memory::DEVICE);
+  }
+
+  /**
+   * @brief Scale a vector by a diagonal matrix in CUDA
+   *
+   * @param[in]  diag - vector representing the diagonal matrix
+   * @param[in, out]  vec - vector to be scaled
+   * @param[in] diag_offset - the index of diag where the diagonal matrix begins
+   *
+   * @pre The diagonal vector must be of the same size as the vector.
+   * @pre vec is unscaled
+   * @post vec is scaled
+   * @invariant diag
+   *
+   * @return 0 if successful, 1 otherwise
+   */
+  void VectorHandlerCuda::scal(vector::Vector* diag, vector::Vector* vec, index_type diag_offset)
+  {
+    real_type* diag_data = &diag->getData(memory::DEVICE)[diag_offset];
     real_type* vec_data  = vec->getData(memory::DEVICE);
     index_type n         = vec->getSize();
     cuda::scale(n, diag_data, vec_data);